Move reshaped_view() to libertem.common.buffers
